envs: tier-2 correctness + comparability cleanup
- Reach tolerance: align UR5e's ``reach_tolerance`` from 0.05 to 0.02
  to match RX200 / NED2 / VX300S; cross-robot success rates are now
  measured against the same radius. (See the matching reach.rst
  updates in the UniROS docs.)
- Float32 dtype: the four goal-reach envs (RX200 + NED2, sim + real)
  now build ``self.ee_pos`` and the concatenated observation as
  ``np.float32`` so the returned obs / achieved-goal actually live in
  the declared ``Box(..., dtype=np.float32)`` space. Removes Gymnasium
  PassiveEnvChecker warnings and a silent 2x-memory cast in the HER
  replay buffer.
- Real reach default loop rate: the six real-reach envs (rx200 / ned2 /
  vx300s, std + goal) defaulted ``environment_loop_rate=None`` then
  divided by it during __init__, so ``gym.make`` with default kwargs
  raised. Default to the hardware rate (RX200 / VX300S 10 Hz, NED2
  25 Hz) so the registry's documented default works.
- Sim/real reach reward parity: the eight real-reach envs defaulted
  ``simple_dense_reward=False`` while their sim counterparts defaulted
  ``True``, so sim and real ran materially different dense rewards for
  the same task. Defaults now match across sim and real.
- Registration hygiene: delete the 5 stray module-level reach
  ``register(...)`` blocks that declared ``max_episode_steps=1000``
  and shadowed the canonical ``__init__.py`` ids (which use 100);
  re-namespace the duplicate ``VX300SRobotGoalEnv-v0`` real
  registration to ``VX300SRobotGoalBaseRealEnv-v0``; fix the
  ``ned1.sim`` entry-point typo in the NED2 sim goal robot env and
  re-namespace its duplicate base id; drop unused
  ``euler_from_matrix`` imports in 4 RX200 robot envs.

+++ src/rl_environments/ned2/sim/task_envs/reach/ned2_reach_goal_sim.py
@@ -425,7 +425,8 @@ class NED2ReacherGoalEnv(ned2_robot_goal_sim.NED2RobotGoalEnv):
 
         # get initial ee pos and joint values (we need this for delta actions or when we have EE action space)
         ee_pos_tmp = self.get_ee_pose()  # Get a geometry_msgs/PoseStamped msg
-        self.ee_pos = np.array([ee_pos_tmp.pose.position.x, ee_pos_tmp.pose.position.y, ee_pos_tmp.pose.position.z])
+        self.ee_pos = np.array([ee_pos_tmp.pose.position.x, ee_pos_tmp.pose.position.y, ee_pos_tmp.pose.position.z],
+                               dtype=np.float32)
         self.ee_ori = np.array([ee_pos_tmp.pose.orientation.x, ee_pos_tmp.pose.orientation.y,
                                 ee_pos_tmp.pose.orientation.z,
                                 ee_pos_tmp.pose.orientation.w])  # for IK calculation - EE actions
@@ -867,7 +868,8 @@ class NED2ReacherGoalEnv(ned2_robot_goal_sim.NED2RobotGoalEnv):
 
         # --- 1. Get EE position
         ee_pos_tmp = self.get_ee_pose()  # Get a geometry_msgs/PoseStamped msg
-        self.ee_pos = np.array([ee_pos_tmp.pose.position.x, ee_pos_tmp.pose.position.y, ee_pos_tmp.pose.position.z])
+        self.ee_pos = np.array([ee_pos_tmp.pose.position.x, ee_pos_tmp.pose.position.y, ee_pos_tmp.pose.position.z],
+                               dtype=np.float32)
         self.ee_ori = np.array([ee_pos_tmp.pose.orientation.x, ee_pos_tmp.pose.orientation.y,
                                 ee_pos_tmp.pose.orientation.z, ee_pos_tmp.pose.orientation.w])
 
@@ -902,9 +904,13 @@ class NED2ReacherGoalEnv(ned2_robot_goal_sim.NED2RobotGoalEnv):
             while not done:
                 done = self._check_joint_states_ready()
 
-        # our observations
+        # our observations. dtype=np.float32 matches the env's
+        # observation_space (Box(..., dtype=np.float32)); without it
+        # numpy promotes to float64 and Gymnasium's PassiveEnvChecker
+        # warns, and SB3 silently casts every step in the HER buffer.
         obs = np.concatenate((self.ee_pos, vec_ee_goal, euclidean_distance_ee_goal, self.joint_pos_all,
-                              prev_action, self.current_joint_velocities), axis=None)
+                              prev_action, self.current_joint_velocities), axis=None,
+                             dtype=np.float32)
 
         if self.log_internal_state:
             rospy.loginfo(f"Observations --->: {obs}")
